Issue #757 : Extend monitoring - added k8s deployment monitoring, removed non-compiling test for now
