+++ deploy/contents/install/install-config
@@ -136,6 +136,9 @@ CP_VM_MONITOR_REQUIRED_LABELS=runid
 CP_VM_MONITOR_RUN_ID_LABEL=Name
 CP_VM_MONITOR_TO_USER=pipe_admin
 CP_VM_MONITOR_CC_USERS=
+CP_VM_MONITOR_PKI_EXPIRATION_DAYS=7
+CP_VM_MONITOR_DEPLOY_HOUR_INTERVAL=1
+CP_VM_MONITOR_DEPLOY_NAMESPACE=default
 
 # cp-share-srv
 CP_SHARE_SRV_INTERNAL_HOST=cp-share-srv.default.svc.cluster.local
