+++ deploy/docker/cp-vm-monitor/config/application.properties
@@ -2,6 +2,7 @@
 cloud.pipeline.host=https://${CP_API_SRV_INTERNAL_HOST:cp-api-srv.default.svc.cluster.local}:${CP_API_SRV_INTERNAL_PORT:31080}/pipeline/restapi/
 cloud.pipeline.token=${CP_API_JWT_ADMIN}
 cloud.pipeline.api.version=0.15
+cloud.pipeline.platform.name=${CP_PREF_UI_PIPELINE_DEPLOYMENT_NAME:Cloud Pipeline}
 
 #Moniroting settings in ms
 monitor.schedule.cron=0 0 */${CP_VM_MONITOR_HOUR_INTERVAL} * * *
@@ -9,12 +10,33 @@ monitor.instance.tag=${CP_VM_MONITOR_INSTANCE_TAG_NAME}=${CP_VM_MONITOR_INSTANCE
 monitor.required.labels=${CP_VM_MONITOR_REQUIRED_LABELS}
 monitor.runid.label=${CP_VM_MONITOR_RUN_ID_LABEL}
 
-#Notification settings
-notification.missing-node.subject=[CloudPipeline] VM $templateParameters["instanceName"] is not registered in cluster
-notification.missing-node.template=file://${CP_VM_MONITOR_TEMPLATES_LOCATION}/MISSING-NODE.html
+#Certificate-monitoring settings
+monitor.cert.schedule.cron=0 0 0 ? * *
+monitor.cert.file.masks=pem
+monitor.cert.expiration.notification.days=${CP_VM_MONITOR_PKI_EXPIRATION_DAYS}
+monitor.cert.scan.folders=${CP_VM_MONITOR_PKI_PATH:},${CP_SHARE_SRV_CERT_DIR:},${CP_EDGE_CERT_DIR:},${CP_DOCKER_CERT_DIR:},${CP_IDP_CERT_DIR:},${CP_GITLAB_CERT_DIR:},${CP_API_SRV_CERT_DIR:},${CP_COMMON_CERT_DIR:}
 
-notification.missing-labels.subject=[CloudPipeline] VM $templateParameters["nodeName"] is missing required labels
-notification.missing-labels.template=file://${CP_VM_MONITOR_TEMPLATES_LOCATION}/templates/MISSING-LABELS.html
+# K8s deployment monitor
+monitor.k8s.deployment.cron=0 */${CP_VM_MONITOR_DEPLOY_HOUR_INTERVAL} * ? * *
+monitor.k8s.deployment.default.namespace=${CP_VM_MONITOR_DEPLOY_NAMESPACE:default}
+monitor.k8s.deployment.names=${CP_VM_MONITOR_DEPLOY_NAMES:cp-git,cp-git-sync,cp-edge,cp-docker-registry,\
+  cp-docker-comp,cp-clair,cp-search-elk,cp-search-srv,cp-heapster-elk,cp-heapster,cp-share-srv,cp-dav}
+
+#Notification settings
+notification.missing-node.subject=[${CP_PREF_UI_PIPELINE_DEPLOYMENT_NAME:Cloud Pipeline}] VM $templateParameters["instanceName"] is not registered in cluster
+notification.missing-node.template=${CP_VM_MONITOR_TEMPLATES:classpath:}/templates/MISSING-NODE.html
+
+notification.missing-labels.subject=[${CP_PREF_UI_PIPELINE_DEPLOYMENT_NAME:Cloud Pipeline}] VM $templateParameters["nodeName"] is missing required labels
+notification.missing-labels.template=${CP_VM_MONITOR_TEMPLATES:classpath:}/templates/MISSING-LABELS.html
+
+notification.cert-expiration.subject=[${CP_PREF_UI_PIPELINE_DEPLOYMENT_NAME:Cloud Pipeline}] Certificate expires soon
+notification.cert-expiration.template=${CP_VM_MONITOR_TEMPLATES:classpath:}/templates/CERTIFICATE-EXPIRATION.html
+
+notification.missing-deploy.subject=[${CP_PREF_UI_PIPELINE_DEPLOYMENT_NAME:Cloud Pipeline}] Missing k8s deployment
+notification.missing-deploy.template=${CP_VM_MONITOR_TEMPLATES:classpath:}/templates/MISSING-DEPLOY.html
+
+notification.not-ready-deploy.subject=[${CP_PREF_UI_PIPELINE_DEPLOYMENT_NAME:Cloud Pipeline}] k8s deployment is not ready
+notification.not-ready-deploy.template=${CP_VM_MONITOR_TEMPLATES:classpath:}/templates/NOT-READY-DEPLOY.html
 
 notification.to-user=${CP_VM_MONITOR_TO_USER}
 notification.copy-users=${CP_VM_MONITOR_CC_USERS}
