+++ vm-monitor/src/main/java/com/epam/pipeline/vmmonitor/service/notification/NotificationConfiguration.java
@@ -1,0 +1,39 @@
+/*
+ * Copyright 2017-2019 EPAM Systems, Inc. (https://www.epam.com/)
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ * http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ *
+ */
+
+package com.epam.pipeline.vmmonitor.service.notification;
+
+import lombok.Data;
+import org.springframework.boot.context.properties.ConfigurationProperties;
+import org.springframework.context.annotation.Configuration;
+
+@Configuration
+@ConfigurationProperties(prefix = "notification")
+@Data
+public class NotificationConfiguration {
+
+    private String missingNodeSubject;
+    private String missingNodeTemplate;
+    private String missingLabelsSubject;
+    private String missingLabelsTemplate;
+    private String certificateExpirationSubject;
+    private String certificateExpirationTemplate;
+    private String missingDeploymentSubject;
+    private String missingDeploymentTemplate;
+    private String notReadyDeploymentSubject;
+    private String notReadyDeploymentTemplate;
+}
