+++ vm-monitor/src/main/java/com/epam/pipeline/vmmonitor/service/notification/VMNotificationService.java
@@ -16,18 +16,12 @@
 
 package com.epam.pipeline.vmmonitor.service;
 
-import com.epam.pipeline.entity.cluster.NodeInstance;
-import com.epam.pipeline.vmmonitor.model.cert.PkiCertificate;
-import com.epam.pipeline.vmmonitor.model.vm.VirtualMachine;
-
-import java.util.List;
+import java.util.Map;
 
 /**
  * Service for building and sending supported types of notifications
  */
 public interface VMNotificationService {
 
-    void notifyMissingNode(VirtualMachine vm);
-    void notifyMissingLabels(VirtualMachine vm, NodeInstance node, List<String> labels);
-    void notifyExpiringCertificate(PkiCertificate certificate);
+    void sendMessage(Map<String, Object> parameters, String subject, String template);
 }
