+++ vm-monitor/src/main/java/com/epam/pipeline/vmmonitor/service/vm/VMMonitorService.java
@@ -12,16 +12,23 @@
  * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
  * See the License for the specific language governing permissions and
  * limitations under the License.
+ *
  */
 
-package com.epam.pipeline.vmmonitor.service;
+package com.epam.pipeline.vmmonitor.service.vm;
 
-import com.epam.pipeline.vo.notification.NotificationMessageVO;
+import com.epam.pipeline.entity.region.AbstractCloudRegion;
+import com.epam.pipeline.entity.region.CloudProvider;
+import com.epam.pipeline.vmmonitor.model.vm.VirtualMachine;
+
+import java.util.List;
 
 /**
- * Service for sending notifications
+ * Cloud provider specific service for getting list of VMs from the cloud
+ * @param <T> supported region
  */
-public interface NotificationSender {
+public interface VMMonitorService<T extends AbstractCloudRegion> {
 
-    void sendMessage(NotificationMessageVO messageVO);
+    List<VirtualMachine> fetchRunningVms(T region);
+    CloudProvider provider();
 }
