+++ vm-monitor/src/main/resources/application.properties
@@ -3,6 +3,8 @@ cloud.pipeline.host=
 cloud.pipeline.token=
 cloud.pipeline.api.version=0.15
 
+cloud.pipeline.platform.name=${CP_PREF_UI_PIPELINE_DEPLOYMENT_NAME:Cloud Pipeline}
+
 #VM-monitoring settings
 monitor.schedule.cron=0 */2 * ? * *
 monitor.instance.tag=monitored=true
@@ -13,17 +15,23 @@ monitor.runid.label=Name
 monitor.cert.schedule.cron=0 0 0 ? * *
 monitor.cert.file.masks=pem
 monitor.cert.expiration.notification.days=5
-monitor.cert.scan.folders=${CP_SHARE_SRV_CERT_DIR},${CP_EDGE_CERT_DIR},${CP_DOCKER_CERT_DIR},${CP_IDP_CERT_DIR},${CP_GITLAB_CERT_DIR},${CP_API_SRV_CERT_DIR},${CP_COMMON_CERT_DIR}
+monitor.cert.scan.folders=${CP_VM_MONITOR_PKI_PATH:},${CP_SHARE_SRV_CERT_DIR:},${CP_EDGE_CERT_DIR:},${CP_DOCKER_CERT_DIR:},${CP_IDP_CERT_DIR:},${CP_GITLAB_CERT_DIR:},${CP_API_SRV_CERT_DIR:},${CP_COMMON_CERT_DIR:}
 
 #Notification settings
-notification.missing-node.subject=[CloudPipeline] VM $templateParameters["instanceName"] is not registered in cluster
-notification.missing-node.template=classpath:/templates/MISSING-NODE.html
+notification.missing-node.subject=[${CP_PREF_UI_PIPELINE_DEPLOYMENT_NAME:Cloud Pipeline}] VM $templateParameters["instanceName"] is not registered in cluster
+notification.missing-node.template=${CP_VM_MONITOR_TEMPLATES:classpath:}/templates/MISSING-NODE.html
 
-notification.missing-labels.subject=[CloudPipeline] VM $templateParameters["nodeName"] is missing required labels
-notification.missing-labels.template=classpath:/templates/MISSING-LABELS.html
+notification.missing-labels.subject=[${CP_PREF_UI_PIPELINE_DEPLOYMENT_NAME:Cloud Pipeline}] VM $templateParameters["nodeName"] is missing required labels
+notification.missing-labels.template=${CP_VM_MONITOR_TEMPLATES:classpath:}/templates/MISSING-LABELS.html
 
-notification.cert-expiration.subject=[CloudPipeline] Certificate expires soon
-notification.cert-expiration.template=classpath:/templates/CERTIFICATE-EXPIRATION.html
+notification.cert-expiration.subject=[${CP_PREF_UI_PIPELINE_DEPLOYMENT_NAME:Cloud Pipeline}] Certificate expires soon
+notification.cert-expiration.template=${CP_VM_MONITOR_TEMPLATES:classpath:}/templates/CERTIFICATE-EXPIRATION.html
+
+notification.missing-deploy.subject=[${CP_PREF_UI_PIPELINE_DEPLOYMENT_NAME:Cloud Pipeline}] Missing k8s deployment
+notification.missing-deploy.template=${CP_VM_MONITOR_TEMPLATES:classpath:}/templates/MISSING-DEPLOY.html
+
+notification.not-ready-deploy.subject=[${CP_PREF_UI_PIPELINE_DEPLOYMENT_NAME:Cloud Pipeline}] k8s deployment is not ready
+notification.not-ready-deploy.template=${CP_VM_MONITOR_TEMPLATES:classpath:}/templates/NOT-READY-DEPLOY.html
 
 notification.to-user=TEST
 notification.copy-users=TEST,TEST
