+++ vm-monitor/src/main/resources/templates/MISSING-LABELS.html
@@ -33,9 +33,9 @@
 <p>Dear user,</p>
 <p>*** This is a system generated email, do not reply to this email ***</p>
 <p>Looks like there is a node $templateParameters["nodeName"] in Cloud Pipeline cluster that is missing one of required labels $templateParameters["missingLabels"].</p>
-<p>This node won't be managed automatically by Cloud Pipeline Platform, please verify node state and terminate it manually.</p>
+<p>This node won't be managed automatically by $templateParameters["platformName"] Platform, please verify node state and terminate it manually.</p>
 <p>Best regards,</p>
-<p>Cloud Pipeline Platform</p>
+<p>$templateParameters["platformName"] Platform</p>
 </body>
 
 </html>
